mnist_project/requirements.txt
fastapi
uvicorn
gradio
torch
torchvision
pillow
